+++ app/Http/Controllers/Api/OrganisasiController.php
@@ -45,7 +45,7 @@ class OrganisasiController extends Controller
                 }
                 return response()->json([
                     'success' => true,
-                    'message' => 'OK',
+                    'message' => 'Berhasil',
                     'code'    => 200,
                     'data'    => $result
                 ]);
@@ -88,9 +88,7 @@ class OrganisasiController extends Controller
             else {
                 $validator = Validator::make($this->request->all(), [
                     'title' => 'required',
-                    'slug' => 'required',
                     'content' => 'required',
-                    'status' => 'required',
                 ]);
                 if ($validator->fails()) {
                     return response()->json($validator->messages());
@@ -107,7 +105,7 @@ class OrganisasiController extends Controller
 
                         return response()->json([
                             'success' => true,
-                            'message' => 'OK',
+                            'message' => 'Berhasil',
                             'code'    => 200,
                             'data'    => $result
                         ]);
@@ -134,7 +132,7 @@ class OrganisasiController extends Controller
             }
             return response()->json([
                 'success' => true,
-                'message' => 'OK',
+                'message' => 'Berhasil',
                 'code'    => 200,
                 'data'    => $result
             ]);
@@ -175,7 +173,7 @@ class OrganisasiController extends Controller
                 }
                 return response()->json([
                     'success' => true,
-                    'message' => 'OK',
+                    'message' => 'Berhasil',
                     'code'    => 200,
                     'data'    => $result
                 ]);
@@ -218,31 +216,30 @@ class OrganisasiController extends Controller
                 ]);
             }
             else {
-
+                $result = $this->service->getOne($id);
                 if (!$result) {
                     return response()->json([
                         'success' => false,
-                        'message' => 'Service Error',
-                        'code'    => 500,
+                        'message' => 'Data Organisasi Tidak Ditemukan!',
+                        'code'    => 404,
                     ]);
                 } else {
                     $result = $this->service->update($this->request, $id);
-                    // $result = $this->service->store($this->request);
-                        if (!$result) {
-                            return response()->json([
-                                'success' => false,
-                                'message' => 'Service Error',
-                                'code'    => 500,
-                            ]);
-                        }
-
+                    if (!$result) {
                         return response()->json([
-                            'success' => true,
-                            'message' => 'OK',
-                            'code'    => 200,
-                            'data'    => $result
+                            'success' => false,
+                            'message' => 'Service Error',
+                            'code'    => 500,
                         ]);
                     }
+
+                    return response()->json([
+                        'success' => true,
+                        'message' => 'Berhasil',
+                        'code'    => 200,
+                        'data'    => $result
+                    ]);
+                }
             }
         } catch (\Throwable $th) {
             dd('Controller error ' . $th->getMessage());
@@ -272,11 +269,20 @@ class OrganisasiController extends Controller
                 ]);
             }
             else {
+                $result = $this->service->getOne($id);
+                if (!$result) {
+                    return response()->json([
+                        'success' => false,
+                        'message' => 'Data Organisasi Tidak Ditemukan!',
+                        'code'    => 404,
+                    ]);
+                }
+
                 $result = $this->service->delete($id);
 
                 return response()->json([
                     'success' => true,
-                    'message' => 'OK',
+                    'message' => 'Berhasil',
                     'code'    => 200,
                 ]);
             }
